requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg[binary]==3.2.9
jinja2==3.1.4
python-multipart==0.0.12
itsdangerous==2.2.0
bcrypt==4.0.1
python-dotenv==1.0.1
qrcode[pil]==8.0
pillow==10.4.0
jpholiday==0.1.8
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
aiofiles==24.1.0
python-jose[cryptography]==3.3.0
requests==2.32.3
